cadenas en c++ corregido

+++ 8Desempeño/Numericos.c
@@ -1,10 +1,11 @@
+//gcc Numericos.c -lm -o num1
 #include <math.h>
 #include <stdio.h> 
 
 int main(){
-    int max = 1757600; 
-    float seno,cose,tang,logn,raiz;
-    int i = 0;
+    double max = 1757600.0; 
+    double seno,cose,tang,logn,raiz;
+    double i = 0.0;
     while(i < max){
         seno += sin(i);
         cose += cos(i);
